Use more meaningful labels
This PR enforces the use of more meaningful labels for all the RYR
projects. It does that by leveraging
[labeler](https://github.com/tonglil/labeler)

The Makefile was updated to simplify the setup and applying the new
labels to all the projects.

The file used by old way to do it using `github label manager` were
deleted.

Fixes: #17