Deploy RabbitMQ
Adds the RabbitMQ service to the developer environment. This service
is required to work with Celery tasks.

Due to https://github.com/kubernetes/charts/issues/5479, the service is
currently configured to run as "ClusterIP" with an ingress controler.
While this works inside of a Kubernetes cluster, it does not work for
local development.

As a result, Redis will be temporarily used as a broker. Once The
issue #5479 will be resolved, RabbitMQ will become the broker.

Fixes request-yo-racks/infra#50