Replace dashboard waiting condition
Due to a security issue, the minikube dashboard is now only available on
demand and the `minikube dashboard` command does not exit anymore. This
command was used to detect when minikube was ready to accept other
deployments.

This patch replaces this command with anither one checking whether the
dashboard service is up and running.

Fixes request-yo-racks/infra#59